Break out stats script class
